requirements-dev.txt
pytest==7.4.4
pytest-mock==3.12.0
pytest-httpx==0.28.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pre-commit==3.6.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
types-docutils==0.20.0.20240106
types-requests==2.31.0.20240106
types-setuptools==69.0.0.20240115
types-urllib3==1.26.25.14
typing_extensions==4.9.0
tox==4.12.1
